@@ -2,11 +2,11 @@ import React from "react";
 //Css
 import "./App.css";
 //Router
-import { Switch, Route, BrowserRouter } from "react-router-dom";
+import { Routes, Route, BrowserRouter } from "react-router-dom";
 //Screens
 import Acceuil from "./screens/Acceuil";
 import Contact from "./screens/Contact";
-import Cours_Tarifs from "./screens/Cours_Tarifs";
+import CoursTarifs from "./screens/CoursTarifs";
 import LaSalle from "./screens/LaSalle";
 import Parcours from "./screens/Parcours";
 //Components
@@ -18,13 +18,13 @@ const App = () => {
     <div className="App">
       <BrowserRouter>
         <Navbar />
-        <Switch>
-          <Route exact path="/" component={Acceuil} />
-          <Route path="/cours_tarifs" component={Cours_Tarifs} />
-          <Route path="/lasalle" component={LaSalle} />
-          <Route path="/contact" component={Contact} />
-          <Route path="/parcours" component={Parcours} />
-        </Switch>
+        <Routes>
+          <Route exact path="/" element={<Acceuil/>} />
+          <Route path="/cours_tarifs" element={<CoursTarifs/>} />
+          <Route path="/lasalle" element={<LaSalle/>} />
+          <Route path="/contact" element={<Contact/>} />
+          <Route path="/parcours" element={<Parcours/>} />
+        </Routes>
         <Footer />
       </BrowserRouter>
     </div>
